Adapted commands to new pipeline runner API
Commands now takes at least 3 arguments:
* Current event
* Current pipeline
* Current context

This was a necessary change to keep the PipelineRunner API clear and
concise. Before that, one had to go to the pipeline's context to search
for sub pipelines, which was ugly and problematic for dynamic pipeline
injection (e.g. macro).

This is now solved.